+++ src/Transmitron/Widgets/Edit.cpp
@@ -53,7 +53,7 @@ Edit::Edit(
     -1,
     "",
     wxDefaultPosition,
-    wxSize(optionsHeight, optionsHeight)
+    wxSize((int)optionsHeight, (int)optionsHeight)
   );
   mSaveSnippet->SetBitmap(wxArtProvider::GetBitmap(wxART_FILE_SAVE));
   mSaveSnippet->Bind(wxEVT_BUTTON, [this](wxCommandEvent &){
@@ -76,8 +76,8 @@ Edit::Edit(
   mTop = new wxBoxSizer(wxOrientation::wxHORIZONTAL);
   mVsizer = new wxBoxSizer(wxOrientation::wxVERTICAL);
   mBottom = new wxBoxSizer(wxOrientation::wxHORIZONTAL);
-  mTop->SetMinSize(0, mOptionsHeight);
-  mBottom->SetMinSize(0, mOptionsHeight);
+  mTop->SetMinSize(0, (int)mOptionsHeight);
+  mBottom->SetMinSize(0, (int)mOptionsHeight);
 
   mRetainedFalse = new wxStaticBitmap(this, -1, *bin2c_not_pinned_18x18_png);
   mRetainedFalse->SetToolTip("Not retained");
@@ -110,7 +110,7 @@ Edit::Edit(
     -1,
     *bin2c_send_18x18_png,
     wxDefaultPosition,
-    wxSize(mOptionsHeight, mOptionsHeight)
+    wxSize((int)mOptionsHeight, (int)mOptionsHeight)
   );
   mPublish->Bind(wxEVT_BUTTON, [this](wxCommandEvent &){
     auto e = new Events::Edit(Events::EDIT_PUBLISH);
